requirements.txt
# Cleaned requirements for Vercel deployment
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
websockets
psycopg2-binary
pandas
numpy
scikit-learn
scipy
joblib
gunicorn
openai
bcrypt
